pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "hdx-scraper-buildings"
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = [
    "duckdb",
    "geopandas",
    "hdx-python-api",
    "hdx-python-country",
    "hdx-python-utilities",
    "httpx[http2]",
    "pandas",
    "pyarrow",
    "python-dotenv",
    "quantulum3[classifier]",
    "tenacity",
    "tqdm",
]

[dependency-groups]
dev = ["pre-commit", "pytest", "pytest-cov", "ruff", "taskipy"]

[tool.hatch]
version.source = "uv-dynamic-versioning"
build.hooks.version.path = "src/hdx/scraper/buildings/_version.py"
build.targets.wheel.packages = ["src/hdx"]

[tool.pytest.ini_options]
addopts = "--color=yes"
log_cli = true
pythonpath = "src"

[tool.ruff]
exclude = ["_version.py"]
fix = true
lint.select = ["ALL"]
lint.ignore = [
    "COM812",
    "D100",
    "D104",
    "D203",
    "D213",
    "INP001",
    "S603",
    "S607",
    "S608",
    "TID252",
]

[tool.taskipy.tasks]
# uv run task app
app = "python run.py"
google = "python -m hdx.scraper.buildings.google"
microsoft = "python -m hdx.scraper.buildings.microsoft"
ruff = "ruff format && ruff check && ruff format"
